requirements.dev.txt
black
isort
flake8
mypy

django-stubs
pandas-stubs

pytest
pytest-django